# swagger.json
{
  "info": {
    "openapi": "3.0.0",
    "title": "SQL-Learning-MK3 test",
    "version": "1.0.0",
    "description": "Testing for SQL-Learning-MK3_backend"
  },
  "tags": [
    {
      "name": "User",
      "description": "ユーザー関連"
    },
    {
      "name": "Auth",
      "description": "認証"
    },
    {
      "name": "Question",
      "description": "問題"
    }
  ],
  "apis": [
    "./dist/index.js"
  ],
  "swagger": "2.0",
  "paths": {
    "/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "ログイン",
        "description": "Login & Get jwt",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Google auth id_token",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "id_token": {
                  "type": "string",
                  "description": "Google認証のid_token"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "headers": {
              "Set-Cookie": {
                "description": "Access token by jwt",
                "schema": {
                  "type": "string",
                  "example": "token=hogehoge"
                }
              },
              "token": {
                "description": "フロント側でのユーザー情報取得用jwt",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/questions": {
      "get": {
        "tags": [
          "Question"
        ],
        "summary": "問題一覧取得",
        "description": "Get questions list",
        "responses": {
          "200": {
            "description": "Success",
            "content": "application/json",
            "schema": {
              "$ref": "#/definitions/QuestionList"
            }
          },
          "404": {
            "description": "Not Found Error",
            "content": "application/json",
            "schema": {
              "$ref": "#/definitions/NotFoundError"
            }
          }
        }
      }
    },
    "/questions/:id": {
      "get": {
        "tag": [
          "Question"
        ],
        "summary": "問題詳細取得",
        "description": "Get question info",
        "responses": {
          "200": {
            "description": "Success",
            "content": "application/json",
            "schema": {
              "$ref": "#/definitions/Question"
            }
          },
          "404": {
            "description": "Not Found Error",
            "content": "application/json",
            "schema": {
              "$ref": "#/definitions/NotFoundError"
            }
          }
        }
      }
    },
    "/users": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "ユーザー一覧取得",
        "description": "Get users list.",
        "responses": {
          "200": {
            "description": "Success",
            "content": "application/json",
            "schema": {
              "type": "object",
              "properties": {
                "users": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/User"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "User"
        ],
        "summary": "ユーザー追加",
        "description": "Add User",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Add User Object",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "email": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/user/:id": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "ユーザー詳細取得",
        "description": "Get user info",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "user id to fetch",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": "application/json",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "400": {
            "description": "Resource not found error",
            "content": "application/json"
          }
        }
      }
    },
    "/users/:id": {
      "put": {
        "201": {
          "description": "Success"
        },
        "400": {
          "description": "Bad request"
        },
        "500": {
          "description": "Internal Server Error"
        },
        "tags": [
          "User"
        ],
        "summary": "ユーザ情報修正",
        "description": "Update User",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "User Object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": null
      }
    }
  },
  "definitions": {
    "NotFoundError": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "description": "Error request"
        },
        "name": {
          "type": "string",
          "description": "Error name"
        }
      }
    },
    "Question": {
      "type": "object",
      "properties": {
        "ID": {
          "type": "integer"
        },
        "title": {
          "type": "string"
        },
        "classID": {
          "type": "integer"
        },
        "sqliteBase64": {
          "type": "string",
          "description": "base64 encoded sqliteFile"
        }
      }
    },
    "QuestionForList": {
      "type": "object",
      "properties": {
        "ID": {
          "type": "integer"
        },
        "title": {
          "type": "string"
        }
      }
    },
    "QuestionList": {
      "type": "object",
      "properties": {
        "questions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/QuestionForList"
          }
        }
      }
    },
    "User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "email": {
          "type": "string"
        },
        "is_Admin": {
          "type": "boolean"
        }
      }
    }
  },
  "responses": {},
  "parameters": {},
  "securityDefinitions": {}
}